requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
attrs==24.2.0
frozenlist==1.5.0
idna==3.10
iso8601==2.1.0
multidict==6.1.0
propcache==0.2.0
py-cord==2.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.2.1
python-dotenv==1.0.1
pytz==2024.2
tortoise-orm==0.21.7
typing_extensions==4.12.2
yarl==1.17.1
pillow==11.0.0
